refactor(winners): 우승자 확인하는 과정을 Winners에게 위임
