+++ src/main/java/racingcar/game/RacingGame.java
@@ -6,19 +6,17 @@ import java.util.List;
 import racingcar.car.Car;
 
 public class RacingGame {
-	private static final int ZERO = 0;
-
-	private List<Car> carList;
+	private List<Car> cars;
 	private Movement movement;
 
-	public RacingGame(List<Car> carList) {
-		this.carList = carList;
+	public RacingGame(List<Car> cars) {
+		this.cars = cars;
 		this.movement = new Movement();
 	}
 
 	public List<String> race() {
 		List<String> racingResult = new ArrayList<>();
-		for (Car each : carList) {
+		for (Car each : cars) {
 			moveCar(each);
 			addResult(racingResult, each);
 		}
@@ -35,26 +33,8 @@ public class RacingGame {
 		racingResult.add(car.positionToString());
 	}
 
-	public List<String> findFrontCars() {
-		List<String> frontCarsName = new ArrayList<>();
-		int maxDistance = getMaxDistance();
-		for (Car each : carList) {
-			insertCarName(frontCarsName, each, maxDistance);
-		}
-		return frontCarsName;
-	}
-
-	private int getMaxDistance() {
-		int maxDistance = ZERO;
-		for (Car each : carList) {
-			maxDistance = Math.max(maxDistance, each.getPosition());
-		}
-		return maxDistance;
-	}
-
-	private void insertCarName(List<String> frontCarsName, Car car, int maxDistance) {
-		if (car.getPosition() == maxDistance) {
-			frontCarsName.add(car.getName());
-		}
+	public String getWinners() {
+		Winners winners = new Winners(cars);
+		return winners.getWinnersMessage();
 	}
 }
